implement CandidType for str (#234)
